backend/requirements.txt
django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
pygithub==1.57
pyjwt==2.6.0
python-telegram-bot==13.11
requests==2.28.1
